Apply context to UnknownStepException (#187)
