+++ tests/DataProvider/Factory/Test/MalformedPageElementReferenceDataProviderTrait.php
@@ -1,0 +1,174 @@
+<?php
+
+namespace webignition\BasilParser\Tests\DataProvider\Factory\Test;
+
+use webignition\BasilParser\Builder\StepBuilder;
+use webignition\BasilParser\Exception\MalformedPageElementReferenceException;
+use webignition\BasilParser\Factory\StepFactory;
+use webignition\BasilParser\Factory\Test\ConfigurationFactory;
+use webignition\BasilParser\Factory\Test\TestFactory;
+use webignition\BasilParser\Model\ExceptionContext\ExceptionContext;
+use webignition\BasilParser\Model\ExceptionContext\ExceptionContextInterface;
+use webignition\BasilParser\Tests\Services\FixturePathFinder;
+
+trait MalformedPageElementReferenceDataProviderTrait
+{
+    public function createFromTestDataThrowsMalformedPageElementReferenceExceptionDataProvider(): array
+    {
+        // MalformedPageElementReferenceException
+        //   thrown when trying to uses a page element reference that is not of the correct form
+        //
+        //   cases:
+        //   - assertion string contains malformed reference
+        //   - action string contains malformed reference
+        //   - test.elements contains malformed reference
+
+        return [
+            'MalformedPageElementReferenceException: assertion string contains malformed reference (1)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    'step name' => [
+                        StepFactory::KEY_ASSERTIONS => [
+                            'malformed_reference is "assertion one value"',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "malformed_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step name',
+                    ExceptionContextInterface::KEY_CONTENT => 'malformed_reference is "assertion one value"',
+                ]),
+            ],
+            'MalformedPageElementReferenceException: assertion string contains malformed reference (2)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    'step name' => [
+                        StepFactory::KEY_ASSERTIONS => [
+                            '".heading" is "assertion one value"',
+                            'malformed_reference is "assertion two value"',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "malformed_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step name',
+                    ExceptionContextInterface::KEY_CONTENT => 'malformed_reference is "assertion two value"',
+                ]),
+            ],
+            'MalformedPageElementReferenceException: action string contains malformed reference (1)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    'step name' => [
+                        StepFactory::KEY_ACTIONS => [
+                            'click action_one_element_reference',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "action_one_element_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step name',
+                    ExceptionContextInterface::KEY_CONTENT => 'click action_one_element_reference',
+                ]),
+            ],
+            'MalformedPageElementReferenceException: action string contains malformed reference (2)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    'step name' => [
+                        StepFactory::KEY_ACTIONS => [
+                            'click ".heading"',
+                            'click action_two_element_reference',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "action_two_element_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step name',
+                    ExceptionContextInterface::KEY_CONTENT => 'click action_two_element_reference',
+                ]),
+            ],
+            'MalformedPageElementReferenceException: test.elements contains malformed reference (1)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    TestFactory::KEY_IMPORTS => [
+                        TestFactory::KEY_IMPORTS_STEPS => [
+                            'step_import_name' => FixturePathFinder::find('Step/element-parameters.yml'),
+                        ],
+                        TestFactory::KEY_IMPORTS_PAGES => [],
+                    ],
+                    'step one' => [
+                        StepBuilder::KEY_USE => 'step_import_name',
+                        StepBuilder::KEY_ELEMENTS => [
+                            'heading' => 'invalid_page_element_reference',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "invalid_page_element_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step one',
+                ]),
+            ],
+            'MalformedPageElementReferenceException: test.elements contains malformed reference (2)' => [
+                'name' => 'test name',
+                'testData' => [
+                    TestFactory::KEY_CONFIGURATION => [
+                        ConfigurationFactory::KEY_BROWSER => 'chrome',
+                        ConfigurationFactory::KEY_URL => 'http://example.com',
+                    ],
+                    TestFactory::KEY_IMPORTS => [
+                        TestFactory::KEY_IMPORTS_STEPS => [
+                            'step_import_name' => FixturePathFinder::find('Step/element-parameters.yml'),
+                        ],
+                        TestFactory::KEY_IMPORTS_PAGES => [],
+                    ],
+                    'step one' => [
+                        StepFactory::KEY_ASSERTIONS => [
+                            '$page.url is "http://example.com"',
+                        ],
+                    ],
+                    'step two' => [
+                        StepBuilder::KEY_USE => 'step_import_name',
+                        StepBuilder::KEY_ELEMENTS => [
+                            'heading' => 'malformed_page_element_reference',
+                        ],
+                    ],
+                ],
+                'expectedException' => MalformedPageElementReferenceException::class,
+                'expectedExceptionMessage' => 'Malformed page element reference "malformed_page_element_reference"',
+                'expectedExceptionContext' =>  new ExceptionContext([
+                    ExceptionContextInterface::KEY_TEST_NAME => 'test name',
+                    ExceptionContextInterface::KEY_STEP_NAME => 'step two',
+                ])
+            ],
+        ];
+    }
+}
